Removed params in preperation for merging configs
We currently used two different `JSON` files to configure the model. One contains information directly relevant to the DeepSpeed engine, while the other contains information about the model, the dataset, and other shit too probably. This is annoying and poor design, so we are going to combine the JSON files.